Move CLI parsing to `grease-cli` package
The test-suite needs access to the CLI to support the `flags:`
directive. However, it is unfortunate to have optparse-applicative as a
dependency of grease (the library). To improve the situation, move all
parsing-related code into the `grease-cli` package.

Fixes #119.
